requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
sqlmodel==0.0.8
python-dotenv==1.0.0
openai==1.98.0
numpy==1.24.3
pydantic==1.10.13
typing-extensions==4.12.2
httpx==0.27.2
pytest==7.4.4
psycopg2-binary==2.9.9

# Phase 1: Vector DB & Embeddings
pgvector==0.2.4
sentence-transformers==2.3.1
faiss-cpu==1.7.4
lightgbm==4.6.0
umap-learn==0.5.5
joblib==1.3.2

# Caching
redis==5.0.1

# Authentication
PyJWT==2.8.0

# Menu Ingestion (PDF, Image, Web)
pdfplumber==0.11.9
pytesseract==0.3.13
pillow==12.1.0
beautifulsoup4==4.14.3
playwright==1.58.0
aiofiles==25.1.0
